requirements.txt
openai
pandas
streamlit
unstructured
tiktoken
pinecone-client
langchain
yfinance
wikipedia
yahooquery
faiss-cpu
python-dotenv
pypdf2